OA user type
