+++ lib/school/exam_arrange/page/index.dart
@@ -24,8 +24,8 @@ class _ExamArrangePageState extends State<ExamArrangePage> {
   late SemesterInfo initial = () {
     final now = DateTime.now();
     return (
-    year: now.month >= 9 ? now.year : now.year - 1,
-    semester: now.month >= 3 && now.month <= 7 ? Semester.term2 : Semester.term1,
+      year: now.month >= 9 ? now.year : now.year - 1,
+      semester: now.month >= 3 && now.month <= 7 ? Semester.term2 : Semester.term1,
     );
   }();
   late SemesterInfo selected = initial;
